refactor(widget_audit): extract function and fix LCOM4 by splitting into modules
Two refactorings applied to src/ui/widget_audit.rs:

1. Extract Function: ControlInventory::new() into 10 private builder functions
   - toolbar_controls() - 7 tools
   - properties_panel_controls() - 5 fields
   - alignment_panel_controls() - 8 controls
   - style_panel_stroke_controls() - 3 controls
   - style_panel_fill_controls() - 3 controls
   - layers_panel_controls() - 5 controls
   - history_panel_controls() - 1 control
   - character_panel_controls() - 6 controls (Phase 2)
   - paragraph_panel_controls() - 3 controls (Phase 2)
   - canvas_text_editor_controls() - 1 control (Phase 2)

   new() now orchestrates these builders with .extend() calls.
   Removed #[expect(clippy::too_many_lines)] from new().
   Added #[expect(clippy::too_many_lines)] to 4 builder functions
   that exceed 70 lines (toolbar, properties, alignment, layers, character).

2. Fix LCOM4 (Low Cohesion): Split into module directory
   - src/ui/widget_audit/types.rs - Phase, ControlSurface, RequirementSource enums
   - src/ui/widget_audit/control.rs - RequiredControl and metadata structs
   - src/ui/widget_audit/inventory.rs - ControlInventory with all builder functions
   - src/ui/widget_audit/mod.rs - Re-exports for backward compatibility

   Public API remains identical - all existing use paths continue to work.

Tests:
✓ cargo check --workspace passes
✓ 25 unit tests pass (widget_audit_test)
✓ 5 shell seam tests pass (gpui_widget_audit_shell_seam)

No behavior changes - pure refactoring for code organization and maintainability.
